graph/vision_node_utils/requirements.txt
gradio

# LangChain
langchain-core
langchain-google-vertexai
pillow
pre-commit
pydantic
python-dotenv
